LayoutTests/http/tests/wasm/wasm-esm-disabled-with-setting-expected.txt

CONSOLE MESSAGE: TypeError: 'application/wasm' is not a valid JavaScript MIME type.
Test that Wasm ESM integration is disabled when flag is off.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.type === 'error' is true
PASS successfullyParsed is true

TEST COMPLETE
